extract constants and functions to utils
There are several constants that are reused throughout the app that are
distinct enough to be defined in their own utilities file. Similarly
there are two functions which need not be defined in the Game component.
