+++ src/Game.tsx
@@ -1,18 +1,15 @@
 import { useState } from "react";
 import Board from "./board/Board";
-
-type Player = "X" | "O";
-export type Move = Player | undefined;
-
-export const cellColumns = ["a", "b", "c"] as const;
-type CellColumn = typeof cellColumns[number];
-export const cellRows = ["3", "2", "1"] as const;
-type CellRow = typeof cellRows[number];
-
-export type CellCoord = { row: CellRow; column: CellColumn };
-
-type RowState = Record<CellColumn, Move>;
-export type BoardState = Record<CellRow, RowState>;
+import {
+  BoardState,
+  calculateWinner,
+  cellColumns,
+  CellCoord,
+  cellRows,
+  getMoveNotation,
+  Player,
+  RowState,
+} from "./utils";
 
 const Game = () => {
   const newBoardState = (boardState?: BoardState) =>
@@ -38,38 +35,7 @@ const Game = () => {
   ]);
   const [stepNumber, setStepNumber] = useState(0);
 
-  const nextPlayer = stepNumber % 2 === 0 ? "X" : "O";
-
-  const calculateWinner = (boardState: BoardState) => {
-    const lines = [
-      ...cellRows.map((row) => cellColumns.map((column) => ({ row, column }))),
-      ...cellColumns.map((column) => cellRows.map((row) => ({ row, column }))),
-      // Generate diagonal lines
-      cellRows.map((row, index) => ({ row, column: cellColumns[index] })),
-      cellRows.map((row, index) => ({
-        row,
-        column: cellColumns.slice().reverse()[index],
-      })),
-    ];
-
-    for (const line of lines) {
-      const winner = boardState[line[0].row][line[0].column];
-      if (
-        winner &&
-        line.every((coord) => boardState[coord.row][coord.column] === winner)
-      ) {
-        return winner as Player;
-      }
-    }
-
-    return null;
-  };
-
-  const getMoveNotation = (
-    player: Player,
-    { row, column }: CellCoord,
-    isWinner: boolean
-  ) => `${player}${column}${row}${isWinner ? "#" : ""}`;
+  const nextPlayer: Player = stepNumber % 2 === 0 ? "X" : "O";
 
   const handleBoardClick = ({ row, column }: CellCoord) => {
     const historySlice = history.slice(0, stepNumber + 1);
